scrapers/requirements.txt
httpx>=0.27.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
schedule>=1.2.0
pydantic>=2.6.0
tenacity>=8.2.0
